re-structure RestFitness.kt and RestResourceFitness.kt to reduce duplicated codes
