+++ core/src/main/kotlin/org/evomaster/core/problem/rest/service/RestFitness.kt
@@ -68,75 +68,10 @@ open class RestFitness : AbstractRestFitness<RestIndividual>() {
             }
         }
 
-        if(actionResults.any { it is RestCallResult && it.getTcpProblem() }){
-            /*
-                If there are socket issues, we avoid trying to compute any coverage.
-                The caller might restart the SUT and try again.
-                Hopefully, this should be just a glitch...
-                TODO if we see this happening often, we need to find a proper solution.
-                For example, we could re-run the test, and see if this one always fails,
-                while others in the archive do pass.
-                It could be handled specially in the archive.
-             */
-            return null
-        }
-
-
-        val ids = targetsToEvaluate(targets, individual)
-
-        val dto = rc.getTestResults(ids)
-        if (dto == null) {
-            log.warn("Cannot retrieve coverage")
-            return null
-        }
-
-        dto.targets.forEach { t ->
-
-            if (t.descriptiveId != null) {
-
-                val noMethodReplacement = !config.useMethodReplacement && t.descriptiveId.startsWith(ObjectiveNaming.METHOD_REPLACEMENT)
-                val noNonIntegerReplacement = !config.useNonIntegerReplacement && t.descriptiveId.startsWith(ObjectiveNaming.NUMERIC_COMPARISON)
-
-                if (noMethodReplacement || noNonIntegerReplacement) {
-                    return@forEach
-                }
-
-                idMapper.addMapping(t.id, t.descriptiveId)
-            }
-
-            fv.updateTarget(t.id, t.value, t.actionIndex)
-        }
-
-        handleExtra(dto, fv)
-
-        handleResponseTargets(fv, individual.seeActions(), actionResults, dto.additionalInfoList)
-
-        if (config.expandRestIndividuals) {
-            expandIndividual(individual, dto.additionalInfoList, actionResults)
-        }
-
-        if (config.baseTaintAnalysisProbability > 0) {
-            assert(actionResults.size == dto.additionalInfoList.size)
-            //TODO add taint analysis for resource-based solution
-            TaintAnalysis.doTaintAnalysis(individual, dto.additionalInfoList, randomness)
-        }
+        restActionResultHandling(individual, targets, actionResults, fv)?:return null
 
         return EvaluatedIndividual(fv, individual.copy() as RestIndividual, actionResults, trackOperator = individual.trackOperator, index = time.evaluatedIndividuals, config = config)
     }
-
-    private fun registerNewAction(action: RestAction, index: Int){
-
-        rc.registerNewAction(ActionDto().apply {
-            this.index = index
-            //for now, we only include specialized regex
-            this.inputVariables = action.seeGenes()
-                    .flatMap { it.flatView() }
-                    .filterIsInstance<StringGene>()
-                    .filter { it.getSpecializationGene() != null && it.getSpecializationGene() is RegexGene}
-                    .map { it.getSpecializationGene()!!.getValueAsRawString()}
-        })
-    }
-
 
 
     override fun doInitializingActions(ind: RestIndividual) {
